refactor(play): extract validateSetupPlacement helper (S26.0e)
Cinquieme tranche du refactor S26.0 :
play/[id]/page.tsx 1348 -> 1322 lignes (-26, total -344 depuis 1666).

- Extrait l'appel API POST /match/:id/validate-setup en fonction
  pure validateSetupPlacement() dans
  apps/web/app/play/[id]/utils/validate-setup.ts.
- handleValidatePlacement (page.tsx) garde la responsabilite UI :
  setSetupSubmitting + try/catch + showSetupError finally.
- L'helper hydrate les 3 etats React via callbacks
  (setState, setIsMyTurn, setMyTeamSide) passes en options.

Slices restantes pour S26.0 : SetupPhaseUI grosso modo (handleDrop +
onCellClick + reserve handling), BlockChoice/ThrowTeamMate flows,
types Move unions pour eliminer les `as any` (13 occurrences).
